Fix potential memory violation and leak in sauer's vector::pop
Before this commit, vector::pop would return an lvalue
reference to the last element in the vector and just mark
that as unused.
This was a terrible idea for a number of reasons.

* If that last element contains an Object, it's destructor
  is never called. I assume the author was under the
  impression that the caller would delete the lvalue, but
  an lvalues destructor is never called (except manually).
* If the lvalue continues to be used, and a new
  element is pushed to the vector, we have a memory
  corruption
* If the lvalue continues to be used and the vector/backing
  buffer is deleted we have a use after free scenario
  (SEGFAULT)

So now we no longer return an lvalue and use std::move to
forward ownership over the object to the caller.